deliveryOptimizator/requirements.txt
googlemaps
numpy
ratelimit
ortools
pandas
python-dotenv
shapely
